Fixed: Out of order commits lead to exception.
The previous revision in the Git log is not neseccarily the parent,
which led to exception since the log was processed sequentially and the
previous commit was used for diffing. This fix makes use of the special
"^" syntax to let Git automatically determine the parent commit for
diffing.

This will lead to problems with other VCS systems and needs a proper fix
later.