Fixed: Missing revisions in analysis due to reflection issue.
By now, we silently ate the ReflectionException which was thrown when
StaticReflection could not find a related class of one we analyzed (e.g.
PHPUnit_Framework_TestCase). The new NullSourceResolver and default
session setup makes StaticReflections simply return a
NullReflectionClass for each of these, which does not hurt since we do
not analyzes related classes anyway.